+++ src/app.module.ts
@@ -9,11 +9,11 @@ import { ConfigModule } from "./modules/config/config.module";
 import { DatabaseModule } from "./modules/database/database.module";
 import { GamesModule } from "./modules/games/games.module";
 import { GarbageCollectionModule } from "./modules/garbage-collection/garbage-collection.module";
-import { HealthModule } from "./modules/health/health.module";
 import { MediaModule } from "./modules/media/media.module";
 import { MetadataModule } from "./modules/metadata/metadata.module";
 import { ProgressModule } from "./modules/progresses/progress.module";
 import { SavefileModule } from "./modules/savefiles/savefile.module";
+import { StatusModule } from "./modules/status/status.module";
 import { UsersModule } from "./modules/users/users.module";
 
 @Module({
@@ -31,7 +31,7 @@ import { UsersModule } from "./modules/users/users.module";
     ScheduleModule.forRoot(),
     EventEmitterModule.forRoot(),
     GarbageCollectionModule,
-    HealthModule,
+    StatusModule,
   ],
   providers: [
     {