+++ src/modules/admin/admin.module.ts
@@ -2,11 +2,11 @@ import { Module } from "@nestjs/common";
 
 import { DatabaseModule } from "../database/database.module";
 import { DatabaseService } from "../database/database.service";
-import { HealthModule } from "../health/health.module";
+import { StatusModule } from "../status/status.module";
 import { AdminController } from "./admin.controller";
 
 @Module({
-  imports: [HealthModule, DatabaseModule],
+  imports: [StatusModule, DatabaseModule],
   controllers: [AdminController],
   providers: [DatabaseService],
 })