+++ src/modules/status/models/authentication-method.enum.ts
@@ -1,0 +1,4 @@
+export enum AuthenticationMethod {
+  BASIC_AUTH = "basic",
+  OAUTH2 = "oauth2",
+}