+++ src/modules/status/models/status.enum.ts
@@ -1,4 +1,4 @@
-export enum HealthStatus {
+export enum StatusEnum {
   HEALTHY = "HEALTHY",
   UNHEALTHY = "UNHEALTHY",
 }